Fix issues with "now" in rules
"now" values would be slightly off because we
would keep regenerating it from Date.now().
Once when replacing the values in the data
sent to the server and again when checking 
the rules. This change sets the now value once
and it is then used consistently in the new data
and in the rules checking.